+++ Questionable/Controller/MiniTaskController.cs
@@ -63,8 +63,10 @@ internal abstract class MiniTaskController<T> : IDisposable
                 catch (Exception e)
                 {
                     _logger.LogError(e, "Failed to start task {TaskName}", upcomingTask.ToString());
+                    var msg = _LF("Failed to start task '{0}'", upcomingTask);
                     _chatGui.PrintError(
-                        $"无法启动任务 '{upcomingTask}', 请使用 /xllog 来获取报错信息", CommandHandler.MessageTag, CommandHandler.TagColor);
+                        _LF("{0} Please check /xllog for more details.", msg), CommandHandler.MessageTag, CommandHandler.TagColor);
+                    _serviceProvider.GetRequiredService<NotificationMasterIpc>().NotifyOnFailure(msg);
                     Stop("Task failed to start");
                     return;
                 }
@@ -88,9 +90,10 @@ internal abstract class MiniTaskController<T> : IDisposable
         {
             _logger.LogError(e, "Failed to update task {TaskName}",
                 _taskQueue.CurrentTaskExecutor.CurrentTask.ToString());
+            var msg = _LF("Could not complete '{0}': {}.", _taskQueue.CurrentTaskExecutor.CurrentTask, e.Message);
             _chatGui.PrintError(
-                $"无法更新任务 '{_taskQueue.CurrentTaskExecutor.CurrentTask}', 请使用 /xllog 来获取报错信息.", CommandHandler.MessageTag, CommandHandler.TagColor);
-
+                _LF("{0} Please check /xllog for more details.", msg), CommandHandler.MessageTag, CommandHandler.TagColor);
+            _serviceProvider.GetRequiredService<NotificationMasterIpc>().NotifyOnFailure(msg);
             Stop("Task failed to update");
             return;
         }