+++ Questionable/QuestionablePlugin.cs
@@ -151,7 +151,6 @@ public sealed class QuestionablePlugin : IDalamudPlugin
         serviceCollection.AddSingleton<StylistIpc>();
         serviceCollection.AddSingleton<MogmailIpc>();
         serviceCollection.AddSingleton<RotationSolverRebornIpc>();
-        serviceCollection.AddSingleton<ItemVendorLocationIpc>();
 
         serviceCollection.AddSingleton<GearStatsCalculator>();
         serviceCollection.AddSingleton<DailyRoutinesIpc>();
@@ -310,7 +309,6 @@ public sealed class QuestionablePlugin : IDalamudPlugin
         serviceCollection.AddSingleton<InterruptHandler>();
 
         serviceCollection.AddSingleton<HighlightObject>();
-        serviceCollection.AddSingleton<HighlightMenus>();
         serviceCollection.AddSingleton<PartyWatchDog>();
 
         serviceCollection.AddSingleton<CraftworksSupplyController>();