+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- bug fix: I broke the skip button, but it's fixed now -alydev
-- Feature: add button to Quick Access to clear VBM obstacle maps -alydev
-  - for if you notice your character running off into the sunset during a combat step for no apparent reason
+- bug fix: "could not retrieve next quest step" is not always a failure state, removing message until better understood -alydev